Update continuous integration system (#54)
* Add library.properties to "Compile Examples" workflow path filter

The library.properties metadata file content is a factor in compilation, so the workflow should run on any change to this
file.

* Use major version refs of sketch compilation actions

Previously, due to the lack of a release, the development versions of the sketch compilation actions were used. Using
release versions provides a more stable CI system for the ArduinoCore-mbed project.

Use of the major version ref will cause the workflow to benefit from ongoing development to the actions up until such
time as a new major release of an action is made, at which time we would need to evaluate whether any changes to the
workflow are required by the breaking change that triggered the major release before updating the major ref
(e.g., uses: 	`arduino/compile-sketches@v2`).

* Optimize management of size deltas report when there are errors

There may sometimes be failed compilations that are not related to the current pull request (e.g., breakage caused by an
external change that will be fixed separately). In this case, it is still valuable to publish a size deltas report for
the compilations that passed.

* Configure Dependabot to check for outdated actions used in workflows

Dependabot will periodically check the versions of all actions used in the repository's workflows. If any are found to
be outdated, it will submit a pull request to update them.
NOTE: Dependabot's PRs will sometimes try to pin to the patch version of the action (e.g., updating `uses: foo/bar@v1`
to `uses: foo/bar@v2.3.4`). When the action author has provided a major version ref, use that instead
(e.g., `uses: foo/bar@v2`). Dependabot will automatically close its PR once the workflow has been updated.
More information:
https://docs.github.com/en/github/administering-a-repository/keeping-your-actions-up-to-date-with-dependabot

* Add badge to readme for "Compile Examples" workflow

This will show the current status of the CI workflow at a glance.

* Add CI workflow to check for commonly misspelled words

On every push, pull request, and periodically, use the codespell-project/actions-codespell action to check for commonly
misspelled words.

In the event of a false positive, the problematic word should be added, in all lowercase, to the ignore-words-list field
of ./.codespellrc. Regardless of the case of the word in the false positive, it must be in all lowercase in the ignore
list. The ignore list is comma-separated with no spaces.

* Correct typos in comments and documentation

* Add CI workflow to do Arduino project-specific linting

On every push, pull request, and periodically, run Arduino Lint to check for common problems not related to the project
code.

* Add license file

This defines the license in a standardized place, which allows GitHub's automated license detection system to detect the
license type:
https://docs.github.com/en/github/creating-cloning-and-archiving-repositories/licensing-a-repository#detecting-a-license

This is the exact license text provided by choosealicense.com
This text should not be modified in any way. Doing so may interfere with GitHub's license detection or even the licence's
function as a legal instrument.

* Use release version of boards platform in "Compile Examples" workflow

Previously, the development version of the boards platform was used to compile the example sketches in the CI workflow.