Reduce Arduino Lint compliance level to "specification"
Linting the library in "strict" compliance mode fails due to the `name` value exceeding the recommended maximum of 16 characters.

It is not possible to change the name, so this is not something that can be fixed. The expected error might mask real problems that cause other rule violations in the future. So the best thing is to reduce the compliance level to "specification" in which the library name length rule violation only results in a warning.